changes required for part 5
to be squashed

+++ part4/blog/controllers/blogs.js
@@ -26,11 +26,10 @@ blogsRouter.post("/", middleware.userExtractor, async (req, res, next) => {
 
   const userId = req.user
 
-
   if (!userId)
     return res.status(400).json({error: 'unable to find userId'})
 
-  const user = await User.findById(userId)
+  let user = await User.findById(userId)
 
   const blog = new Blog({
     title: req.body.title,
@@ -45,7 +44,9 @@ blogsRouter.post("/", middleware.userExtractor, async (req, res, next) => {
     user.blogs = [result._id]
   else 
     user.blogs = user.blogs.concat(result.id)
-  await user.save()
+  user = await user.save()
+
+  result.user = user
 
   res.status(201).json(result);
 });
@@ -53,20 +54,30 @@ blogsRouter.post("/", middleware.userExtractor, async (req, res, next) => {
 blogsRouter.delete("/:id", middleware.userExtractor, async (req, res, next) => {
   const userId = req.user
   const blogToDelete = await Blog.findById(req.params.id)
+  const user = await User.findById(userId)
+
+  if (!blogToDelete)
+    return res.status(404).json({error: 'blog not found'})
 
   if (userId.toString() !== blogToDelete.user.toString())
     return res.status(403).json({error: 'forbidden action'})
 
   await Blog.findByIdAndDelete(blogToDelete.id)
+  user.blogs = user.blogs.filter(bId => bId.toString() !== req.params.id)
+  // console.log(user.blogs)
+  await user.save()
+
   res.status(204).end()
 });
 
-blogsRouter.put("/:id", async (req, res, next) => {
-  const { title, author, url, likes } = req.body;
+blogsRouter.put("/:id", middleware.userExtractor, async (req, res, next) => {
+  const { title, author, url, likes} = req.body;
 
-  let blogToUpdate = await Blog.findById(req.params.id);
+  let blogToUpdate = await Blog.findById(req.params.id)
+                            .populate("user", {username: 1, name: 1, _id: 1});
 
-  if (!blogToUpdate) return res.status(404).end();
+  if (!blogToUpdate) 
+    return res.status(404).end();
 
   blogToUpdate.title = title ?? blogToUpdate.title;
   blogToUpdate.author = author ?? blogToUpdate.author;
